+++ conductor/src/duration.rs
@@ -4,14 +4,14 @@ use crate::tempo::Tempo;
 #[derive(Copy, Clone, Debug)]
 pub enum Duration {
 	/// Represents a duration of time in seconds.
-	Seconds(f32),
+	Seconds(f64),
 	/// Represents a duration of time in beats.
-	Beats(f32),
+	Beats(f64),
 }
 
 impl Duration {
 	/// Gets the time in seconds.
-	pub fn in_seconds(&self, tempo: Tempo) -> f32 {
+	pub fn in_seconds(&self, tempo: Tempo) -> f64 {
 		match self {
 			Duration::Seconds(seconds) => *seconds,
 			Duration::Beats(beats) => tempo.beats_to_seconds(*beats),
